@@ -5,7 +5,8 @@
   "author": "Michael Davis",
   "bugs": "https://github.com/Contentstack-Solutions/contentstack-cli-content-type/issues",
   "dependencies": {
-    "@contentstack/cli-command": "0.1.0-beta",
+    "@contentstack/cli-command": "^1.2.8",
+    "@contentstack/cli-utilities": "^1.4.4",
     "@oclif/command": "^1.8.0",
     "@oclif/config": "^1.17.0",
     "@types/diff2html": "^3.0.0",
@@ -18,26 +19,24 @@
     "diff2html": "^3.1.17",
     "git-diff": "^2.0.6",
     "moment": "^2.29.1",
-    "node-graphviz": "^0.1.0",
-    "table": "^6.0.4",
+    "node-graphviz": "^0.1.1",
+    "table": "^6.8.1",
     "tmp": "^0.2.1",
     "tslib": "^1.13.0",
     "url-join": "^4.0.1"
   },
   "devDependencies": {
-    "@oclif/dev-cli": "^1.22.2",
     "@oclif/plugin-help": "^3.2.0",
-    "@types/jest": "^26.0.15",
-    "@types/node": "^10.17.28",
+    "@types/jest": "^29.1.2",
+    "@types/node": "^10.17.60",
     "eslint": "^5.16.0",
     "eslint-config-oclif": "^3.1.0",
     "eslint-config-oclif-typescript": "^0.1.0",
     "globby": "^10.0.2",
-    "jest": "^26.6.1",
-    "request": "^2.88.2",
-    "ts-jest": "^26.4.2",
-    "ts-node": "^8.10.2",
-    "typescript": "^3.9.7"
+    "jest": "^29.3.1",
+    "ts-jest": "^29.0.3",
+    "ts-node": "^10.9.1",
+    "typescript": "^4.8.4"
   },
   "engines": {
     "node": ">=8.0.0"
@@ -67,7 +66,7 @@
     "test": "jest",
     "postpack": "rm -f oclif.manifest.json",
     "posttest": "eslint . --ext .ts --config .eslintrc",
-    "prepack": "rm -rf lib && tsc -b && oclif-dev manifest && oclif-dev readme",
-    "version": "oclif-dev readme && git add README.md"
+    "prepack": "rm -rf lib && tsc -b && oclif manifest && oclif readme",
+    "version": "oclif readme && git add README.md"
   }
 }
