@@ -1,87 +1,130 @@
-import Command from '../../core/command'
-import {flags} from '@contentstack/cli-command'
-import cli from 'cli-ux'
-import buildOutput from '../../core/content-type/compare'
+import Command from "../../core/command";
+import {
+  flags,
+  FlagInput,
+  managementSDKClient,
+  cliux,
+  printFlagDeprecation
+} from "@contentstack/cli-utilities";
+import buildOutput from "../../core/content-type/compare";
+import { getStack, getContentType } from "../../utils/index";
 
 export default class CompareCommand extends Command {
-  static description = 'compare two Content Type versions';
+  static description = "Compare two Content Type versions";
 
   static examples = [
-    '$ csdx content-type:compare -s "xxxxxxxxxxxxxxxxxxx" -c "home_page"',
-    '$ csdx content-type:compare -s "xxxxxxxxxxxxxxxxxxx" -c "home_page" -l # -r #',
+    '$ csdx content-type:compare -k "xxxxxxxxxxxxxxxxxxx" -c "home_page"',
+    '$ csdx content-type:compare -k "xxxxxxxxxxxxxxxxxxx" -c "home_page" -l # -r #',
     '$ csdx content-type:compare -a "management token" -c "home_page" -l # -r #',
   ];
 
-  static flags = {
+  static flags: FlagInput = {
     stack: flags.string({
-      char: 's',
-      description: 'Stack UID',
-      required: false,
-      exclusive: ['token-alias'],
+      char: "s",
+      description: "Stack UID",
+      exclusive: ["token-alias"],
+      parse: printFlagDeprecation(['-s', '--stack'], ['-k', '--stack-api-key']),
     }),
 
-    'token-alias': flags.string({
+    "stack-api-key": flags.string({
+      char: "k",
+      description: "Stack API Key",
+      exclusive: ["token-alias"]
+    }),
+
+    "token-alias": flags.string({
+      char: "a",
+      description: "Management token alias",
+      parse: printFlagDeprecation(['--token-alias'], ['-a', '--alias']),
+    }),
+
+    alias: flags.string({
       char: 'a',
-      description: 'management token alias',
-      hidden: false,
-      multiple: false,
-      required: false,
+      description: 'Alias of the management token',
     }),
 
-    'content-type': flags.string({
-      char: 'c',
-      description: 'Content Type UID',
+    "content-type": flags.string({
+      char: "c",
+      description: "Content Type UID",
       required: true,
     }),
 
     left: flags.integer({
-      char: 'l',
-      description: 'Content Type version, i.e. prev version',
+      char: "l",
+      description: "Content Type version, i.e. prev version",
       required: false,
-      dependsOn: ['right'],
+      dependsOn: ["right"],
     }),
 
     right: flags.integer({
-      char: 'r',
-      description: 'Content Type version, i.e. later version',
+      char: "r",
+      description: "Content Type version, i.e. later version",
       required: false,
-      dependsOn: ['left'],
+      dependsOn: ["left"],
     }),
-  }
+  };
 
   async run() {
     try {
-      const {flags} = this.parse(CompareCommand)
-      this.setup(flags)
-
-      cli.action.start(Command.RequestDataMessage)
+      const { flags } = await this.parse(CompareCommand);
+      this.setup(flags);
+      this.contentTypeManagementClient = await managementSDKClient({
+        host: this.cmaHost,
+      });
 
       if (!flags.left) {
-        const discovery = await this.client.getContentType(this.apiKey, flags['content-type'], false)
-        const version = discovery.content_type._version
+        const spinner1 = cliux.loaderV2(Command.RequestDataMessage);
+        const discovery = await getContentType({
+          managementSdk:this.contentTypeManagementClient,
+          apiKey:this.apiKey,
+          uid:flags["content-type"],
+          spinner:spinner1
+        }
+        );
+        const version = discovery._version;
 
-        flags.left = version
-        flags.right = version > 1 ? version - 1 : version
-
-        this.log(`Comparing versions: ${flags.left} <-> ${flags.right}.`)
+        flags.left = version;
+        flags.right = version > 1 ? version - 1 : version;
+        cliux.loaderV2("", spinner1);
+        this.log(`Comparing versions: ${flags.left} <-> ${flags.right}.`);
       }
+
+      const spinner = cliux.loaderV2(Command.RequestDataMessage);
 
       const [stack, previous, current] = await Promise.all([
-        this.client.getStack(this.apiKey),
-        this.client.getContentType(this.apiKey, flags['content-type'], true, flags.left),
-        this.client.getContentType(this.apiKey, flags['content-type'], true, flags.right),
-      ])
+        getStack(this.contentTypeManagementClient, this.apiKey, spinner),
+        getContentType({
+          managementSdk: this.contentTypeManagementClient,
+          apiKey: this.apiKey,
+          uid: flags["content-type"],
+          ctVersion: flags.left,
+          spinner,
+        }),
+        getContentType({
+          managementSdk: this.contentTypeManagementClient,
+          apiKey: this.apiKey,
+          uid: flags["content-type"],
+          ctVersion: flags.right,
+          spinner
+        }),
+      ]);
 
-      cli.action.stop()
+      cliux.loaderV2("", spinner);
 
-      const output = await buildOutput(flags['content-type'], previous, current)
-      this.printOutput(output, 'changes', flags['content-type'], stack.name)
+      const output = await buildOutput(
+        flags["content-type"],
+        previous,
+        current
+      );
+      this.printOutput(output, "changes", flags["content-type"], stack.name);
 
       if (flags.left && flags.left === flags.right) {
-        this.warn('Comparing the same version does not produce useful results.')
+        this.warn(
+          "Comparing the same version does not produce useful results."
+        );
       }
-    } catch (error) {
-      this.error(error, {exit: 1, suggestions: error.suggestions})
+    } catch (error: any) {
+      this.error(error, { exit: 1, suggestions: error.suggestions });
     }
   }
 }
