@@ -1,5 +1,6 @@
 import {Command} from '@contentstack/cli-command'
-import {BuildOutput} from './content-type/build-output'
+import {configHandler, cliux} from '@contentstack/cli-utilities'
+import {BuildOutput} from '../types'
 import ContentstackClient from './contentstack/client'
 
 export default class ContentTypeCommand extends Command {
@@ -9,33 +10,38 @@ export default class ContentTypeCommand extends Command {
 
   protected client!: ContentstackClient;
 
+  protected contentTypeManagementClient: any;
+
   setup(flags: any) {
-    if (!this.authToken) {
+    const authToken = configHandler.get('authtoken');
+    if (!authToken) {
       this.error('You need to login, first. See: auth:login --help', {exit: 2, suggestions: ['https://www.contentstack.com/docs/developers/cli/authentication/']})
     }
+    const stackAPIKey = flags.stack || flags["stack-api-key"]
+    const mTokenAlias = flags['token-alias'] || flags.alias
 
-    if (!flags['token-alias'] && !flags.stack) {
-      this.error('You must provide either a token alias or a Stack UID.', {exit: 2})
+    if (!mTokenAlias && !stackAPIKey) {
+      cliux.print('Error: You must provide either a token alias or a Stack UID.', {"color": "red"});
+      process.exit(1);
     }
 
-    if (flags['token-alias']) {
-      const token = this.getToken(flags['token-alias'])
+    if (mTokenAlias) {
+      const token = this.getToken(mTokenAlias)
 
       if (token.type !== 'management') {
-        this.warn('Possibly using a delivery token. You may not be able to connect to your Stack. Please use a management token.')
+        cliux.print('Possibly using a delivery token. You may not be able to connect to your Stack. Please use a management token.', {"color": "yellow"})
       }
 
       this.apiKey = token.apiKey
     } else {
-      this.apiKey = flags.stack as string
+      this.apiKey = flags.stackAPIKey as string
     }
 
-    this.client = new ContentstackClient(this.cmaHost, this.authToken)
+    this.client = new ContentstackClient(this.cmaHost, authToken)
   }
 
   printOutput(output: BuildOutput, who: string, what: string | null, where: string) {
     this.log(`Requested ${who}${what ? ` for '${what}' ` : ' '}on '${where}.'`)
-    this.log('---\n')
 
     if (output.hasResults) {
       if (output.header) {
@@ -52,5 +58,9 @@ export default class ContentTypeCommand extends Command {
     } else {
       this.log(`No ${who} found.`)
     }
+  }
+
+  async run() {
+
   }
 }
