@@ -1,8 +1,6 @@
-import * as tableImport from 'table'
-import {BuildOutput} from './build-output'
+import {table} from 'table'
+import {BuildOutput} from '../../types'
 import * as format from './formatting'
-
-const {table} = tableImport
 
 export default function buildOutput(logs: any, users: any): BuildOutput {
   const rows = []
